updated user rating, fixed components, done clean code

+++ co-ride-backend/src/main/kotlin/mk/edu/ukim/finki/coride/api/request/RatingRequest.kt
@@ -2,5 +2,5 @@ package mk.edu.ukim.finki.coride.api.request
 
 data class RatingRequest(
         val driverId: Long,
-        val score: Long
+        val score: Double
 )
