+++ co-ride-backend/src/main/resources/db/migration/V4__create_ratings_table.sql
@@ -2,6 +2,6 @@ create table ratings
 (
     id        bigserial primary key,
     driver_id bigint references users (id),
-    score     bigint,
+    score     double precision,
     rider_id  bigint references users (id)
 )
